refactor: exhibition -> event로 네이밍 변경
