+++ src/main/java/com/example/codebase/domain/Event/crawling/dto/eventDetailResponse/XmlEventDetailResponse.java
@@ -1,4 +1,4 @@
-package com.example.codebase.domain.exhibition.crawling.dto.detailExhbitionResponse;
+package com.example.codebase.domain.Event.crawling.dto.eventDetailResponse;
 
 import jakarta.xml.bind.JAXBContext;
 import jakarta.xml.bind.Unmarshaller;
@@ -16,24 +16,24 @@ import java.io.StringReader;
 @XmlRootElement(name = "response")
 @XmlAccessorType(value = XmlAccessType.FIELD)
 @XmlType(propOrder = {"msgBody", "comMsgHeader"})
-public class XmlDetailExhibitionResponse {
+public class XmlEventDetailResponse {
     private MsgBody msgBody;
     private ComMsgHeader comMsgHeader;
 
-    public static XmlDetailExhibitionResponse parse(String body)  {
+    public static XmlEventDetailResponse parse(String body)  {
         try {
-            JAXBContext jaxbContext = JAXBContext.newInstance(XmlDetailExhibitionResponse.class);
+            JAXBContext jaxbContext = JAXBContext.newInstance(XmlEventDetailResponse.class);
             Unmarshaller unmarshaller = jaxbContext.createUnmarshaller();
 
             StringReader reader = new StringReader(body);
-            return (XmlDetailExhibitionResponse) unmarshaller.unmarshal(reader);
+            return (XmlEventDetailResponse) unmarshaller.unmarshal(reader);
 
         } catch (Exception e) {
             throw new RuntimeException(e);
         }
     }
 
-    public XmlDetailExhibitionData getDetailExhibitionData() {
+    public XmlEventDetailData getDetailExhibitionData() {
         return msgBody.getDetailExhibitionData();
     }
 }
